Move logger test utils to relevant package
