+++ packages/logger/package.json
@@ -71,7 +71,8 @@
   "devDependencies": {
     "@types/triple-beam": "^1.3.2",
     "rimraf": "^4.4.1",
-    "triple-beam": "^1.3.0"
+    "triple-beam": "^1.3.0",
+    "@lodestar/test-utils": "^1.10.0"
   },
   "keywords": [
     "ethereum",
